seperate interface and implement
